DS-1080: improve GraphQL error handling
Turns out most of the stuff mentioned at:

https://www.apollographql.com/docs/apollo-server/features/errors/

...is not really relevant to us simply wanting better winston logging
output. It's more focused on filtering out or modifying errors based
on where you want to send them. It also has some information on
using apollo's graph manager, which we may want to think about at
some point but is way out of scope on this task.

Anyway, instead I cleaned up the error reporting code to extract the
stacktrace in a way that's less likely to change and also to just print
out the rest of the full error object in case there's something useful
we missed. Why not just print out that error object? Well, most often
the thing we care about is the stacktrace and the error message, so
we still sort of want that on top and easy to find.