fix issues with broken document downloads
If a document dataset is computed (unstructured to unstructured), it's
not going to have an 'original' path, so we set that to the 'imported'
path. If a document dataset is already raw text (and it's not computed),
passing it through tika is only going to garble it. So we want to
detect that and not do it. However, that means that we won't have a real
imported version in that case. So in that case, we set the imported
path to be the original path. We're also checking the mimetype during
upload to be able to more intelligently deal with our files.