Perform database migrations after deploying
Because the format of our database may change between versions, we
need to perform a migration after each deploy.

We also need to make sure the neo4j database is running before we
attempt to perform the migration. If that doesn't happen within the
timeout (currently approximately 50 seconds), it's a good sign the
deployment failed.