DatasetDetails: Add error boundary
Establish an error boundary to catch errors from any of the child
dataset components and add a new error placeholder comoponent to be
displayed in those cases. Only the tab that actually has the error will
show the error placeholder and the user will still be freely able to
switch to other tabs.