clean up dataset and transformation resolver code
This also moves the util.js code that used to be in resolvers to
domain/context. The general theme is to leave code that's to do with
translating from GraphQL structures and concerns in the resolvers,
bringing the pure business logic into a context function. Contexts are
being borrowed from elixir/phoenix, where they refer to a higher layer
between code representing the database mapping and controllers. Given
the heavy use of context in resolver parameters, there may be better
names.