upgrade react, apollo-client, and material-ui
There are a few warnings that remain when doing an npm install and in
dev mode in the browser... peer dependencies for:

- @date-io/date-fns
- material-ui-chip-input
- material-ui-pickers
- material-ui-search-bar
- react-rangeslider

We won't be able to resolve the last until plot.ly updates the chart
editor to something later. material-ui-search-bar may have to be
brought in or forked to eliminate the warning. It seems to be only a
problem with the dependency declaration, not any actual dependency.
The others may be resolvable, but material-ui-pickers does seem to be
at a state in development where it's struggling to maintain
compatibility with the newer material-ui code. At this point it
definitely works for what we're using it for.

Note that you may have to blow away your node_modules and/or
package-lock.json for this.