fix issues with eslint upgrade
I noticed that eslint wasn't working after some of the upgrades. I
suspect that this is due to eslint 6 itself and different ways in which
it tries to access plugins. In particular, it seems to have a big
problem trying to inherit settings from a parent folder where there's
no actual node project (with the modules installed that that parent
config would need).

I wasn't able to find a place that comes right out and said the way
we were doing things was once supported and now isn't, but reading
links like the following:

https://eslint.org/docs/user-guide/migrating-to-6.0.0#-plugins-and-shareable-configs-are-no-longer-affected-by-eslints-location
https://github.com/eslint/eslint/issues/11720

make me wonder if we were relying on accidental behaviour to begin
with. In any event, the extra configuration on .vscode's eslint plugin
plus moving to specific .eslintrc.js files for both the client and the
server causes eslint to again operate properly. Given the fuzziness
around the behaviour we were exploiting, it seems better to do this
small amount of duplication rather than roll back the package and
avoid the inevitable for a while longer.