Testenv: switch to using pytest to run our test
This will still allow a test that fails due to a python exception to
be caught, but for the moment doesn't enforce that test outcomes are
actually correct.

Note that since (for the moment) we're only testing by manually
reviewing the output of the tests, we're explicitly disabling pytest's
output capturing.