Make tests explicit about what success means
Instead of outputting content, assert that the content is what it
should be. In order to facilitate that, I've replaced the iris
datasets and transformations with very simple datasets and
transformations that have a very small set of results, so we can
succinctly state the full state of what we expect.