add isMember and isOwner rules for graphql-shield
These may end up being our only rules for quite some time. They can be
composed with and/or/not functions from graphql-shield. They're
slightly more complex than the basic graphql-shield rules, where we
basically export factory functions that create the rule instead of just
the rule function. This allows a fieldMap to be given, which basically
maps certain parameters our rule needs to do its checks to the actual
arguments in the query/mutation or fields in the parent object. In
many cases we can use defaults (for example, 'uuid' is a fairly standard
argument/field name). But where it's not as standard, we can map our
names.