replace the rest of directive auth and then some
This should fully replace the need for directive-based auth in the
schema (I've actually removed it completely now). In addition, all
known types/queries/mutations that are tied to an organization and
should require a member to access have been locked down with isMember,
isOwner, or isCurrentUser.

I was also able to remove a lot of cases where the user had to be
passed to lower business logic layers just to verify access permissions.

It should now be fairly easy to lock down new functionality.

I know I've done it in the wrong order, but I'm now going to actually
add some tests to verify various permission scenarios.