change OrganizationID to OrganizationRef
I lied. At least one thing in between the last checkin and adding tests:

For cases where we allow looking up something by name or uuid (or for
what should now be deprecated use, id), we use a GraphQL type instead
of just a raw scalar value. OrganizationID was the first case where
that was tried. But after expanding, Ref seemed to be a better
indication of the intention behind these types (a name, after all,
isn't really an ID, but a reference captures the idea of a pointer to a
unique thing). And that's how OrganizationID became OrganizationRef.

Note that we'll have to update the dependent Python API to reflect
this. But since it's actually another refactoring task this sprint to
turn it into a real pip package, I'll probably just take that on to do
next and do the necessary updates while I'm at it.