sketch out some permissions tests
I'm going to have to redo a lot of this once we've merged Jeremy's test
changes in. I thought about leaving it until then, but didn't want to
think about merging this branch in before adding some automated
verification that we've got at least as good permission protection as
the stuff I ripped out. There are many more scenarios we can and should
check, but I've checked the main ones the old system should have
protected and then some, so I think this is a good enough start. I
don't want to spend too much more time fleshing out these tests until
adopting the real test framework that was added, as the setup and
teardown is starting to get more complex.