don't allow duplicate dataset names
Previously we were enforcing this at the UI level, but we never really
prevented it. Neo4J doesn't allow the type of constraint to prevent it
at that level (though, with some duplication, where we attached
org-uuids to Dataset/Transformation objects, we could create such
constraints if we decide it's important to also enforce them in the
database). This change actively prevents duplicate names for datasets.
Currently the only way a new dataset is created is by eventually calling
createDataset on an organization domain model, so here it does one
extra query that the name isn't already in use before allowing a name
given by whatever's calling the GraphQL query to be used.