update integration tests to use python-adi
python-adi makes testing a fair bit easier because a connection is now
an object instead of a global on the module. I updated existing
non-permission tests by just updating to the new namespacing and method
names. The permission test was more of a challenge due to it not really
being a test yet and having left it a bit messy, anticipating this
update. So this commit also breaks it properly into separate tests with
actual assertions along with updating it to use the new package. And
it's here where the new package really comes in handy, as different user
connections can be set up in advance rather than having to switch
settings back and forth.