add ability to detect encoding while reading
This works slightly differently when using object storage vs legacy
storage, as the object storage methods read everything into memory
first. Yep, that's not smart, and we're going to want to eventually
write files locally before really working with them. But it seems like
a bad idea to not get a few more brains around a whiteboard before
doing that, as pulling down big files to local storage for every little
operation has its own potential problems, and in some cases, we may not
quite need to do that.

Anyway, the result should end up being the same. With object storage, we
attempt to use up to the first 5 MB instead of the whole file to
determine the encoding. This is because the magic library itself seems
much less efficient when reading a set of bytes given to it than when
just given a file path.