write imported versions with utf-8 encoding
I made a small choice here that may have a less than small impact.
Hopefully good and not bad. We could try and keep track of an original
encoding and maintain it throughout all the various reads/writes, and
also provide ways for users to override or access different encodings
when they're interacting with datasets in transformations. But should we
really burden them with that? Of course, we want to leave pure binary
files (such as pdfs) alone. But for csv, text, and text output from
tika, I can't think of a really good reason to pass the complexity of
caring about encodings on further down the line. There are probably some
small efficiencies in some niche cases, and yes, it's going to take up
a little more storage, but well, I'm just going to stop and let better
minds argue for me: https://utf8everywhere.org.

This is why we're only focusing on the import tasks here. We're doing
imports on CSV files anyway, to extract columns and also to normalize
them, so normalizing encoding too is just another step in providing a
solid ground on which to start transforming things. We also ended up,
somewhat accidentally, turning pdf/doc/etc to text conversion into an
import step for non-structured types, instead of requiring a separate
transformation for them. This means that we can do the same for them
so that, structured or unstructured, the 'imported' version of your
data will be in utf-8 format without really having to think about it.