add transformations navigation button
For now, there's just a placeholder component to show the detail screen
for Transformations so that there's not an error when you click the
button.

Note that we're eventually going to have to swap out the left hand bar
when navigating Transformations (and probably Scenarios), as that still
displays datasets. Or we might decide that we want to keep that and
swap it with something else (perhaps filtering / browsing options).