don't read files unnecessarily
In the migration (which we may not even need anymore due to not having
any situations where data would still be in the old form), we were
reading an entire file before checking all the conditions that would
actually mean we need to do anything with it. Now we're not.