test memoizing a method on a model
We probably won't use this. I'm trying to figure out the best way to do
more efficient loading, especially when we often have to do minor data
queries on directly connected nodes (we have to do that a lot with
organizations).

The original thinking was to set a cache that lives only the duration
of the object (as far as I know, refreshed between requests) that we
could map pre-loaded nodes to for efficiency. If they weren't mapped,
they would at least be memoized so they weren't loaded multiple times
to access simple values.

Thinking about this a bit at least made me confident that we should be
able to do something like it if we have to without disrupting too much
other code. But there may be better options if we don't rush. One
example is using the Query object to actually query the related nodes
and organize them in a structure resembling the graph... where we then
initialize our model with that larger graph (instead of just the single
node), and it uses that to initially grab objects, only turning to new
database queries when it runs out of the pre-loaded graph.