allow non-members to see published transformations
Although some of the previous commits opened up published
transformations across organizations, we were still only allowing users
that were members in the organization to see them. We needed to create
another rule to also allow viewing if the transformation is 'published'.

We will probably need to reinvestigate this rule and either add another
one or make it more complex once we start looking at more complex
sharing mechanisms. Right now, all we have to do is check if it's
published, and if it is, we know any logged in user should be able to
see it.