only show publish toggle if you can use it
Users are prevented from actually being able to publish/unpublish a
transformation if they aren't a member of the owning organization.
However, a "Publish" toggle button that just doesn't work is a bit
weird. This commit proactively checks that they could publish it and
doesn't display the toggle when the user doesn't have those permissions.

There's a slight imperfection here from what I initially wanted to do.
To me, it would be more intuitive to hide the "Publish" toggle even on
a transformation where you were a member of the owning organization if
you weren't currently operating under that organization (via the
selector in the top right corner). Right now, you still can. This isn't
a security concern. After all, you could just switch the organization
and make the change even under my ideal solution. But it does blur some
of that distinction over which organization you're operating under,
which I don't think we want. However, to do this, we're going to have
to figure out how to pass more than just the user as context to our
query resolvers. I believe that will be a combination of using a special
ApolloLink instance and setting a header, which we then resolve, much
like the user authorization (though in this case, we can simply verify
that the user belongs to that organization). Then we'd have a
"currentOrg" context item that we could check against in our resolvers.