don't try recreating users when arleady started
Sometimes we might run bin/testenv start because we forgot it was
already up. Sometimes it might just be a quick way to get a service we
know is down up again. However, we don't want to keep trying to create
users and organizations when we do that.

There's also the problem that we were *able* to create duplicate users
and orgs. I'm going to fix that by adding a constraint, but it would
still trigger an error here that we want to avoid.