remove 'virtual' field and compute instead
Right now it costs us an extra query to compute 'virtual' instead of
just set it. But it avoids the abstraction leaking into multiple
places, and we will likely be able to address the extra query more
holistically later (and we have lots of little extra queries we'd like
to address in a similar fashion).