update apoc to 3.4.0.7
This gets a useful apoc.coll.intersection function, among other things,
that the filter can use to check if there's at least one array element
on a neo4j object that matches one of a list of elements passed in.

This will require a rebuild of neo4j for anyone using it. We don't do
that too often. I'm unsure of what to do on staging.