Server: Add ModelPromiseNull type
ModelPromiseNull is the new name for the old ModelPromise type which
allowed null values. The new ModelPromise type no longer allows null values.
This change means that callers now only have to do null checks on return
values that can actually be null. Ideally we eventually convert the rest
of the model functions to throw instead of returning null to remove the
issue completely.