implement collapse/expand suggestion
I initially tried this with the material-ui expander component.
However, that component seems to assume you want to keep all the
contents from the summary section. In our case, we truly have a
"collapsed" view and an "expanded" view, and if we show all of both,
we'll needlessly have duplicate information. Fortunately, material-ui
has a lower level 'Collapse' component that can handle the smooth
animation of removing one area and adding another, and the rest of the
logic isn't too hard to replicate with some simple state.