problem noticed on staging: datasets without type
I fixed this by hand just to verify it was the culprit. We never
officially migrated datasets without a type field, but we increasingly
assume this is always set. It was breaking the new dataset browse
results in some cases. This ensures that datasets without a type will
get defaults of a 'csv' type.