fix error thrown when browsing unowned datasets
When you're not a member of an organization that owns a dataset, you
also don't own the columns or the metadata associated. We were still
being very restrictive with these, so an error was thrown when trying
to access them.

I modified the isPublished() rule to call a method on the object,
rather than just a field. This allows the object to decide how to carry
that operation out. Then we can use the same trick as with canAccess,
where we follow our connections and check if an ancestor has been
published in order to determine if the base object is published.

From a query standpoint, this is very inefficient, but provided we can
address that in a more holistic way ("data loader" techniques or
fetching subgraphs instead of just single nodes... or some other
memoization techniques), it's easier logic to follow.

The other thing, unaddressed here is... it's not the easiest thing in
the world to trace down what a permission error is, and it would
be nice if we simply didn't display the values that couldn't be
accessed instead of just blowing up. I'm leaving that unaddressed
for now, but it would be a good backlog item.