create datasets from reusable transformations
This adds a dialog for creating datasets from reusable transformations.
The dialog allows choosing datasets corresponding to the inputs for the
transformation from the ones the user has access to. The lookup pulls
the first 20, alphabetically, by default. If the user types in anything,
results will be requeried using the string to filter names. This will
be done in a debounced fashion to not send unnecessary graphql calls.