move more permission checking to outer layer
createDataset already had a GraphQL Shield isOwner check. It was doing
the check again in the context function, along with another check on
specific ability to create datasets. Removing these, and creating a new
rule for dataset creation permission allows removing the need to check a
user at this layer.