fix winston mangling an error and confusing us
When we have an instance of a GraphQLError that didn't originate
because of an exception being thrown (intentionally or not) from code
that is core GraphQL implementation (coming directly from apollo-server)
we may end up not having an 'originalError'. Thus we need to check it
before trying to destructure anything from it.