graphql shield: don't swallow errors in dev mode
I'd still rather be able to throw a specific error type that is allowed
to percolate all the way up to the user. And it would be nice to be
able to log the actual errors even in production mode, even if we
don't display them to the user. However, right now it seems like
one or the other, so we can at least default to seeing better errors
in dev mode and avoid pulling our hair out tracking down bugs.