browse: enforce distinctness of filtered datasets
If metadata is added with the same search term as an attribute on the
dataset itself, the query to find datasets based on a search term can
return duplicates. There are many ways to go about fixing this. Trying
to avoid producing those duplicates in the first place would result in
overly complex query logic. Filtering them out after the fact (as we're
now doing in the older search functionality for the dataset workbench)
is needlessly putting off dealing with the duplicates. Because this
newer list functionality is really just one big query, we can use
DISTINCT for what it was meant for.

This change adds the ability to specify that you want the query to only
return distinct results, either upon construction or running of the
query (as with other parameters) and then uses that new option in
listDatasets.