remove artificial metadata separation
This was a tough one. This commit removes the separate node and related
classes/queries for DatasetMetadata. The fields that were there still
exist, but now simply as properties of a Dataset node. There's also a
migration that will move them from the DatasetMetadata nodes to related
Dataset nodes.

I feel this deserves a little more explanation and history, but feel free to
skip the rest if you don't need it.

When this node type came into being, we had a fairly large selection of
fields that we were thinking about adding. We weren't entirely sure of
what they would be used for (they were based off of fields that were
used in a separate data portal). Furthermore, most of them weren't
important to the rest of the current system. They were simply bits of
information that could be set and retrieved. Faced with an unknown
amount of fields that we really just needed to record, and wanting
some separation from fields that were critical to the concepts binding
the rest of ADI together, it felt right to separate these in the
database and throughout the code as well.

Then two slow drifts began to happen. First, we developed more and more
of a model layer around our neo4j components, preferring JS object
interactions, in many cases, over raw queries. This has many advantages,
but one area it makes a little more cumbersome is when you have sets of
nodes that tend to always change together and actually act in a very
coupled manner for the most part. Our model classes deal with a single
type of node at a time, and related nodes tend to be explicitly loaded
in follow up queries.

The other thing that changed gradually is that some of the fields that
were initially added just for bookkeeping actually became more important
for system operations. dateUpdated can help us know if a computed
dataset has become stale. Format is very closely related to the dataset
type. So much so that the type might be better derived from the format,
or at the very least, we'd want to update the type properly to reflect
changes to format (a "csv" type that's a "pdf" format doesn't really
make sense). We could cheat and just expose both fields for the user to
set and/or create very similar fields directly on the Dataset node for
system level issues and make sure to sync those with their metadata
siblings when necessary. And at least right now, we don't want to change
the ability to access all of them on the "Metadata" screen of the
application.

To sum it up, there was a lot of work being done to maintain this
separation, but the separation itself wasn't really giving any benefits.
About the only place it made sense to make the grouping that
DatasetMetadata made was on the "Metadata" UI tab. And nothing about
that screen requires that all the fields in it come from the same place.
In fact, we also include some Column info on it. There are still several
fields that aren't critical to system operation and where we're unsure
of how widely they'll be adopted. At the very least, it's more likely we
may want to separate them back out in some form at some point. I've
marked these with the prefix of "ext_" in our internal fields. It's
really just a marker for development purposes to remind us that we're
not really sure where these belong, but that they probably don't belong
on Dataset. Unfortunately given that many fields from that previous
"metadata" specfic object do now belong (and also could easily be
considered "metadata"), our use of DatasetMetadata for just "fields
we're not sure about" seems silly. What I'm guessing right now is that
as we see more use out of these other fields, we'll either see some not
used at all, and we should get rid of them. Or we'll start to understand
that others are used together for certain purposes and logical groupings
will become more obvious. For example, the fields around "updates" may
make sense to group under a separate node type at some point, especially
if we end up using them to schedule automatic updates of datasets. But
right now, they're just fields that get set. They don't really do
anything.

So now in this brave new world, we don't have to make a separate async
load just to get dateUpdated or format information. These are first class
fields on a dataset.