fix error getting updated info from an import
I'm not sure why we weren't seeing more of this before. Somehow what
I've been doing on this branch brought it out for documents, but it was
always waiting to be a problem. Our import tasks are expecting a "data"
section of the json passed back, and furthermore, "format" and "bytes"
fields in it. If the data section is not there at all (as was the case with the
document imports, there will be an attempt to destructurre undefined. If
the data section is there but the actual data is undefined, a quirk of
our database layer is that the property simply doesn't get set (and
we're left with whatever was originally set). This commit makes sure
we're sending back the expected data in both import tasks.