make more use of the format setting
This extracts functionality to determine the format from a file
extension so that it can be done without being part of an upload
operation.

It also uses format to decide on the ulitimate type of the dataset,
since the two are pretty tightly related. Given the increased importance
of format, we also make sure it has a default value.