let format determine some upload behvaiour
Explictly setting the format from the filename before having the dataset
process the upload (and also reloading the dataset so that it ends up as
the right subclass according to format) means that when the upload
triggers an import, it will trigger the right one (from the right
subclass). As noted, we might want to consider changing the abstraction
a bit now that we're making fewer decisions before the import step.