use last import task status for processing state
Previously, we were using some easy to break assumptions about the flow
of information in determining whether to put the upload buttong into a
processing state and then to keep it spinning. I was breaking the odd
assumption as I was doing this work, and we now actually had the info
on the last task to correctly set the state and have it updated according
to what's in the database (so, for example, you can refresh the browser,
go to the page, and if the dataset's still processing, you'll still see
the spinner... until it actually stops processing).