add ability to query download options of dataset
We usually build download uris from the client side, which is a bit
brittle, and when we end up wanting to allow more than one type
of download, we end up duplicating knowledge between the client
and server. This commit allows the download options to be queried
instead, so we can have the server explicitly tell the client what's
available.