split download button if more than one option
This changes the DownloadButton to look at the downloadOptions from the
returned GraphQL data instead of encoding that knowledge in the client.
It's mainly for presenting a split button when there's more than one
option, but the other benefit is that we let the server supply knowledge
about downloads now instead of duplicating the logic in the client.