fix proptypes issues
Turns out there isn't really a good way to just get the proptypes that
material-ui uses from it, so I copied the potential color strings from
the typescript version and then set className to expect a string.

It sounds like material-ui explicitly discourages you from trying to
access its proptypes, claiming you don't need to when you're wrapping
a component:

https://github.com/mui-org/material-ui/issues/14790

This is a little odd because it's still common to want to pass along
properties to those base components that would be expected to conform
to their proptypes, which is exactly what we're doing here. But there
are certainly more pressing concerns.