update path to get jar file
We should probably come up with a better way to get the most recent
jar or to pin the version. This dist path seems to only contain the most
recent, though it still does it by version number, so you have to know
that, and if they update, it effectively breaks our image build. In any
even, this updates to the latest: tika-app-1.23.jar