make our image ocr actually work again
This seems to trigger the same output as if you used the following
header when posting to a tika-server:

--header "X-Tika-PDFOcrStrategy: ocr_only"

This comes from Tika's documentation on PDFs and OCR:

https://cwiki.apache.org/confluence/display/TIKA/TikaOCR

Unfortunately, they don't make it easy to figure out how to do this when
using tika via tika-app.jar. Perhaps this is something regular Java
developers just take for granted, and therefore they don't explicitly
document it for Tika. But here's the logic I concocted to figure out
the right tika-config.xml setting, and it might help for future settings:

If you go to the following class documentation, you'll find a bunch of
"set" methods:

https://tika.apache.org/1.18/api/org/apache/tika/parser/pdf/PDFParserConfig.html

One of them is setOcrStrategy, with a String parameter. Another is
setExtractInlineImages, with a boolean parameter. We already use this in
our configuration overrides because it was explicitly documented
somewhere online. I expect that any "set" method you find in here (at
least if it has a primitive Java type) can be converted into a parser
param by chopping off the "set", lowercasing the first letter following,
and then putting the new value as the body of the param tag, making
sure to put the type in there (also lowercased first letter) that you
found in the javadocs. This seemed to work for ocrStrategy anyway.

Note: this definitely increases the cost of tika parsing and isn't as
easy to shut off per transformation as it is with setting a server
header. We may end up having to craft our own ability to accept
differing parameters to pass on through.