one step further for proper task dependencies
We're still following a fairly procedural task execution flow,
specifically waiting for one to finish before looking into whether there
are others to follow. But with this change, we can at least kick off
multiple following tasks that depend on a common prerequisite task.

The next step with this could be to only return tasks from task.next()
that do not depend on any other tasks that are unfinished. Eventually
the next() call from one of those other tasks would have those
conditions met.

Even better would be to have some sort of daemon always looking for
tasks that are ready to be run. In that case, it wouldn't really care
whether a preceeding task had been run, only that the current task it
wants to run isn't dependent on anything else.

However, right now, all we really need to be able to do is branch out.