new python-adi transformation bootstrapping
This removes a lot of cruft from the worker that was involved with
loading and running transformations. Now, the job of a worker task
involving a transformation is basically to just find the transform, call
its .run method with the right parameters, and collect any results.

Similarly, registration basically just runs the outer module (like
before) and collects the dataset mappings. Because those
mappings now ultimately have to get to method parameter
names, we have to pass back a bit more information than before.
A non-reusable transformation is basically like a reusable one
with pre-wired defaults now.