update server tasks to handle new worker reqs
We need to pass back slightly more complex information during
registration. We need to pass a bit more info about each dataset
to provide more flexibility in the transformation code. Finally, we
need to store a bit more when making connections between
input datasets and a non-renewable transformation so that
we know which variable names to pass an input dataset to.