no longer need upload override on documentDataset
This was explicitly needed to maintain the 'paths' property of a dataset
model object. Now that paths is recomputed each time via  a getter, we
can get rid of it.

It also didn't have the necessary async and await happening, so it
could have been at the root of some of the weird test timing issues
with Jenkins.