set originalFilename before getting updated paths
Since in some cases the paths will depend on the uploaded filename
(for getting extensions), this has to be set before getting the paths.
Note that there's not a really great reason for sticking the extension
on in object storage, and it does create little gotchas like this. It is
nice for sometimes checking right at the object level to have a better
idea of what the file actually contains, but it might be simpler to do
away with keeping any extensions when storing the file.