Server: Do per request caching of model objects for graphql-shield
Our graphql-shield guards were re-querying each of the relevant model
objects one-by-one because they only had access to the graphql object,
but needed access to the model logic. This resulted in a large number of
small neo4j queries for every request that scaled linearly with the
number of datasets available to the user. Instead we're now caching
those model objects when they're queried by the context layer and the
guards pull them out of the cache to avoid hitting neo4j.