store docker logs after test
Knowing if any errors occurred in the server logs could be especially
helpful when trying to diagnose test failures. I'm removing the inserted
sleeps that allowed tests to pass, so they'll probably fail with this
push. But for the first time (hopefully!) we'll be able to see the
server output under such conditions.