Client: Add ability to view links to datasets
This is hacky and relies on manually parsing the window path,
translating it into navigation state and clearing the path so we don't
get stuck on that dataset. The DatasetView also has to set the
organization in the navigation state based on the owner of the dataset
now so everything shows the correct org when loading a dataset link.
This should NOT be expanded upon and should be removed when we convert
the application to use react-router.